+++ lib/v2/datasource/remote/api/balance_repository.dart
@@ -2,7 +2,7 @@ import 'package:async/async.dart';
 import 'package:http/http.dart' as http;
 import 'package:seeds/v2/datasource/remote/api/network_repository.dart';
 import 'package:seeds/v2/datasource/remote/model/balance_model.dart';
-import 'package:seeds/v2/models/token_model.dart';
+import 'package:seeds/v2/datasource/local/models/token_model.dart';
 export 'package:async/src/result/error.dart';
 export 'package:async/src/result/result.dart';
