+++ lib/v2/screens/dashboard/interactor/usecases/balance_update_use_case.dart
@@ -5,7 +5,7 @@ import 'package:seeds/v2/datasource/remote/api/members_repository.dart';
 import 'package:seeds/v2/datasource/remote/api/planted_repository.dart';
 import 'package:seeds/v2/datasource/remote/api/profile_repository.dart';
 import 'package:seeds/v2/datasource/remote/api/voice_repository.dart';
-import 'package:seeds/v2/models/token_model.dart';
+import 'package:seeds/v2/datasource/local/models/token_model.dart';
 
 export 'package:async/src/result/error.dart';
 export 'package:async/src/result/result.dart';
